Suggest backward compatible, improve clear, remove debug in poll
Suggest now polices old suggestions, but if a user reacts, bot restarts, then user reacts the other, it will be allowed. If the user then unreacts and readds the other one, then it will work as intended again.

+++ commands/moderation/clear.js
@@ -23,7 +23,7 @@ function addDelete(messageID) {
 function deleteSavedMessages(channel) {
     if (markedDelete.length == 0)
         return;
-    if (markedDelete.length == 1)
+    if (markedDelete.length == 1 && channel.messages.cache.find(msg => msg.id == markedDelete[0]).deletable)
         channel.messages.cache.find(msg => msg.id == markedDelete[0]).delete().catch();
     else if (markedDelete.length > 1)
         channel.bulkDelete(markedDelete.slice(0, 100)).catch();
@@ -150,7 +150,8 @@ module.exports = class ClearCommand extends Command {
                                 deleteSavedMessages(msg.channel);
                             else;
                         else {
-                            msg.delete().catch();
+                            if (msg.deletable)
+                                msg.delete().catch();
                         }
                 });
                 deleteSavedMessages(message.channel);
@@ -165,7 +166,8 @@ module.exports = class ClearCommand extends Command {
                                 deleteSavedMessages(msg.channel);
                             else;
                         else {
-                            msg.delete().catch();
+                            if (msg.deletable)
+                                msg.delete().catch();
                         }
                 });
                 deleteSavedMessages(message.channel);
